resize user avatar + contact page
now the avatar loaded by user is resized during storage
+
contact page has been implement with emails setup

+++ app/Http/Controllers/UserController.php
@@ -11,13 +11,15 @@ use Image;
 class UserController extends Controller
 {
     /**
-     * Display a listing of the resource.
+     * Display a listing of all coweerkers.
      *
      * @return \Illuminate\Http\Response
      */
     public function index()
     {
-        //
+        $listOfUsers = User::where([['id', '!=', auth()->id()],['admin', false],])->get();
+
+        return view('coweerkers', ['listOfUsers' => $listOfUsers]);
     }
 
     /**
@@ -55,8 +57,13 @@ class UserController extends Controller
 
         $authUser = User::find(Auth::user()->id);
 
-        Storage::disk('local')->delete('public/'.$authUser->avatar); //delete the old picture
-        $avatar = $request->file('avatar')->store('avatars', 'public');//store the new picture
+        if($authUser->avatar !== 'avatars/defaultAvatar.png') {
+        Storage::disk('local')->delete('public/'.$authUser->avatar); //delete the old picture if different than the default avatar picture
+        };
+
+        $avatar = $request->file('avatar')->store('avatars', 'public'); //store the new picture
+        $image = Image::make(Storage::get('public/'.$avatar))->fit(300)->encode();
+        Storage::put('public/'.$avatar, $image);
 
         $authUser->update(['avatar' => $avatar]); //update the db
 
@@ -110,6 +117,12 @@ class UserController extends Controller
      */
     public function update(Request $request, $id)
     {
+        $this->validate($request, [
+            'name' => 'required|',
+            'job' => 'min:2|max:50',
+            'user_description' => 'min:2|max:200'
+        ]);
+
         $authUser = User::find(Auth::user()->id);
 
         $authUser->name = request('name');
